// src/app/classes_interfaces/interfaces.ts
export interface videoResultInterface {
    adult: boolean
    backdrop_path: string
    genre_ids: Array<number>
    id: number
    original_language: string
    original_title: string
    overview: string
    popularity: any
    poster_path: string
    release_date: string
    title: string
    video: boolean
    vote_average: any
    vote_count: number
}